Use generic list model component at every challenge screen
