+++ challenges/src/main/java/mx/yellowme/youst/challenges/ui/crazylists/CrazyListsChallengeActivity.kt
@@ -1,9 +1,8 @@
-package mx.yellowme.youst.challenges.crazylists
+package mx.yellowme.youst.challenges.ui.crazylists
 
-import androidx.recyclerview.widget.GridLayoutManager
-import androidx.recyclerview.widget.RecyclerView
+import kotlinx.android.synthetic.main.challenge_crazy_lists.*
 import mx.yellowme.youst.challenges.R
-import mx.yellowme.youst.challenges.common.ChallengeWithListActivity
+import mx.yellowme.youst.challenges.common.BaseChallengeActivity
 import mx.yellowme.youst.challenges.domain.ChallengeType
 import mx.yellowme.youst.challenges.domain.Pokemon
 import mx.yellowme.youst.challenges.pokemon.PokemonListLoader
@@ -14,7 +13,7 @@ import mx.yellowme.youst.core.hooks.recycler.ItemListener
 import mx.yellowme.youst.core.hooks.recycler.SimpleRecyclerAdapter
 import java.util.*
 
-class CrazyListsChallengeActivity : ChallengeWithListActivity<Pokemon, PokemonListViewHolder>(),
+class CrazyListsChallengeActivity : BaseChallengeActivity(),
     ItemListener<Pokemon> {
 
     override val layoutId = R.layout.challenge_crazy_lists
@@ -23,21 +22,13 @@ class CrazyListsChallengeActivity : ChallengeWithListActivity<Pokemon, PokemonLi
 
     override fun onResume() {
         super.onResume()
-        PokemonListLoader.loadData(this)
-    }
-
-    override fun initAdapter(): SimpleRecyclerAdapter<Pokemon, PokemonListViewHolder>? {
-        return PokemonListAdapter(ArrayList(), this)
+        PokemonListLoader.loadData(pokeListModel)
     }
 
     override fun onItemClick(item: Pokemon?) {
         item?.name?.let {
             toast(it)
         }
-    }
-
-    override fun initLayoutManager(): RecyclerView.LayoutManager {
-        return GridLayoutManager(this, 3)
     }
 
 }
