+++ challenges/src/main/java/mx/yellowme/youst/challenges/ui/listentome/ListenToMeChallengeActivity.kt
@@ -1,0 +1,13 @@
+package mx.yellowme.youst.challenges.ui.listentome
+
+import mx.yellowme.youst.challenges.R
+import mx.yellowme.youst.challenges.common.BaseChallengeActivity
+import mx.yellowme.youst.challenges.domain.ChallengeType
+
+class ListenToMeChallengeActivity : BaseChallengeActivity() {
+
+    override val layoutId = R.layout.challenge_listen_to_me
+
+    override val currentType = ChallengeType.LISTEN_TO_ME
+
+}
