+++ challenges/src/main/res/layout/challenge_crazy_lists.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="?attr/activityBackground">
+    android:background="?attr/activityBackground"
+    android:orientation="vertical">
 
     <mx.yellowme.youst.core.components.toolbar.MainActionToolbar
         android:id="@+id/mainActionToolbar"
@@ -16,45 +16,9 @@
         app:leftAction="back"
         app:menuAction="@string/help" />
 
-    <TextView
-        android:id="@+id/emptyMessageTextView"
-        style="@style/Challenge.Subtitle"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:visibility="gone"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        tools:text="@string/empty_list_holder" />
-
-    <ProgressBar
-        android:id="@+id/loadingProgressBar"
-        android:layout_width="@dimen/dimen_50"
-        android:layout_height="@dimen/dimen_50"
-        android:indeterminate="true"
-        android:visibility="gone"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/recyclerView"
+    <mx.yellowme.youst.challenges.components.PokeListModel
+        android:id="@+id/pokeListModel"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="@dimen/dimen_12"
-        android:layout_marginEnd="@dimen/dimen_12"
-        android:paddingTop="@dimen/dimen_6"
-        app:layoutManager="androidx.recyclerview.widget.GridLayoutManager"
-        app:layout_constrainedHeight="true"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/mainActionToolbar"
-        app:spanCount="3"
-        tools:itemCount="15"
-        tools:listitem="@layout/card_item_pokemon" />
+        android:layout_height="match_parent" />
 
-</androidx.constraintlayout.widget.ConstraintLayout>
-
+</LinearLayout>
