+++ challenges/src/main/res/layout/component_generic_list_model.xml
@@ -18,6 +18,18 @@
         app:layout_constraintTop_toTopOf="parent"
         tools:text="@string/empty_list_holder" />
 
+    <TextView
+        android:id="@+id/errorMessageTextView"
+        style="@style/Challenge.Subtitle"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="@string/error_list_message"
+        android:visibility="gone"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintRight_toRightOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
     <ProgressBar
         android:id="@+id/loadingProgressBar"
         android:layout_width="@dimen/dimen_50"
